Split-off MetroFramework.Design
This was done to remove the dependency on System.Design.dll, which is
not included with the .NET Framework, Client Profile.

For VS to load the designer classes, these should ideally be registered
in the GAC. However, it is still possible to reference both
MetroFramework.dll and MetroFramework.Design.dll from a project and use
full design-time support. The reference to MetroFramework.Design.dll
should then be dropped by the compiler as it isn't necessary for
run-time support.