MetroTextBox.WithError Property Attribute Fix
Change the `MetroTextBox.WithError` property from [EditorBrowsable(EditorBrowsableState.Never)] to [Browsable(False)]. The "EditorBrowsable" attribute hides this property from appearing in IntelliSense AutoComplete and the Object Browser (unless set to show hidden members) but the WithError property was still being listed in the Visual Studio IDE Properties panel. (I assume the original intent was to hide this property from showing in the Visual Studio Designer? It doesn't really make sense to have the item visible in the Designer Properties window but not appearing as a member in the code IDE IntelliSense...)